Refactor room management and synchronization logic
- Removed RoomContext and RoomLoop implementations to streamline room handling.
- Introduced SafeIDAllocator for thread-safe ID allocation in RoomManager.
- Updated IRoomContext interface to simplify player management by using UID instead of IPlayer interface.
- Enhanced ClientSyncData and ServerSyncData structures for better frame synchronization.
- Added new ClientInputData and WorldEventData types for improved event handling in the game world.
- Implemented CloseWithError method in session interfaces for better error handling.
- Created wire provider set for RoomManager to facilitate dependency injection.
- Added README for sync module to clarify frame synchronization concepts.

+++ src/pkg/lockstep/session/ISession.go
@@ -4,7 +4,7 @@ import "net"
 
 type ISession interface {
 	Close() error
-	CloseWithError(code uint16, reason string) error
+	CloseWithError(code uint32, reason string) error
 	IsConnected() bool
 	SendDatagram(data []byte) error
 	ReceiveDatagram() ([]byte, error)
